# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Dyhs Workspace API 계약 초안
  version: 0.1.0-draft
  description: >-
    전체 목표의 초기 설계 문서입니다. 최신 Live 서버의 실제 계약은 docs/live-api.md를 참조하세요.
    이 초안의 모든 기능과 필드가 구현된 것은 아닙니다. 정적 목업은 이 API를 호출하지 않습니다.
    서버 세션으로 주체를 결정하며 모든 메일·첨부·별칭 접근은 소유권을 검증합니다.
    응답은 Cache-Control no-store, 민감 내용은 로그 금지.
    프론트엔드 관리자 예시 토글은 서버 권한에 사용하지 않습니다.
servers:
  - url: https://workspace.example
security:
  - session: []
paths:
  /api/me:
    get:
      summary: 세션 주체, 연결된 메일함, 서버 권한 및 기능 상태
      responses:
        '200':
          description: 미연결 사용자도 mailbox null로 응답하며 타인 메일함을 대체 표시하지 않음
          content:
            application/json:
              schema:
                type: object
                required: [id, displayName, mailbox, roles, capabilities]
                properties:
                  id: { type: string }
                  displayName: { type: string }
                  mailbox: { type: [string, 'null'], description: 서버에서 검증한 불투명 메일함 ID }
                  roles:
                    type: array
                    items: { enum: [user, mail_operator, platform_admin] }
                  capabilities:
                    type: object
                    additionalProperties: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/mail/folders:
    get:
      summary: 본인 폴더와 읽지 않은 개수
      responses:
        '200':
          description: 연결된 메일함 범위에 한정
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [id, name, unread]
                  properties:
                    id: { type: string }
                    name: { type: string }
                    unread: { type: integer, minimum: 0 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/mail/messages:
    get:
      summary: 서버에서 검색·정렬·페이지 제한을 적용한 목록
      parameters:
        - { name: folder, in: query, required: true, schema: { type: string, maxLength: 256 } }
        - { name: cursor, in: query, schema: { type: string, maxLength: 1024 } }
        - { name: query, in: query, schema: { type: string, maxLength: 200 } }
      responses:
        '200':
          description: 페이지 최대 50개, cursor는 사용자·폴더에 바인딩
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, maxItems: 50, items: { $ref: '#/components/schemas/MessageSummary' } }
                  nextCursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/mail/messages/{id}:
    parameters:
      - { $ref: '#/components/parameters/MessageId' }
    get:
      summary: 메일 본문과 헤더, 첨부 메타데이터
      responses:
        '200':
          description: 서버에서 MIME 디코딩·정화한 본문, 원격 이미지 기본 차단
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/MessageSummary' }
                  - type: object
                    properties:
                      to: { type: array, items: { type: string, format: email } }
                      cc: { type: array, items: { type: string, format: email } }
                      text: { type: string }
                      sanitizedHtml: { type: [string, 'null'] }
                      attachments:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: string }
                            filename: { type: string }
                            size: { type: integer, minimum: 0 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: UIDVALIDITY 변경 등으로 만료된 ID. 목록 갱신 필요 }
  /api/mail/messages/{id}/attachments/{attachmentId}:
    parameters:
      - { $ref: '#/components/parameters/MessageId' }
      - { name: attachmentId, in: path, required: true, schema: { type: string } }
    get:
      summary: 메시지와 첨부의 소유권을 함께 검사한 다운로드
      responses:
        '200':
          description: attachment Content-Disposition(UTF-8 filename*)과 X-Content-Type-Options nosniff 필수
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/mail/messages/{id}/actions:
    parameters:
      - { $ref: '#/components/parameters/MessageId' }
      - { $ref: '#/components/parameters/Csrf' }
    post:
      summary: 소유권과 대상 폴더를 검증한 읽음·이동·휴지통 처리
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              additionalProperties: false
              properties:
                action: { enum: [mark_read, mark_unread, move, trash] }
                targetFolder: { type: string, description: move인 경우 필수; 서버에서 같은 메일함인지 검사 }
      responses:
        '204': { description: 변경 완료 }
        '400': { description: 유효하지 않은 작업 또는 누락된 대상 폴더 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/mail/send:
    post:
      summary: 허용된 발신자·별칭 검증과 중복 방지 후 전송
      description: >-
        SMTP 수락 후 Sent 저장만 실패하면 메일을 다시 전송하지 않습니다.
        통신 단절로 수락 여부를 모르면 delivery_unknown 상태로 제공하고 자동 재발송하지 않습니다.
        첨부 업로드 계약은 실제 서버 제한 확정 후 별도 정의합니다.
      parameters:
        - { $ref: '#/components/parameters/Csrf' }
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [from, to, subject, text]
              properties:
                from: { type: string, format: email }
                to: { type: array, minItems: 1, maxItems: 100, items: { type: string, format: email } }
                cc: { type: array, maxItems: 100, items: { type: string, format: email } }
                subject: { type: string, maxLength: 250 }
                text: { type: string, maxLength: 1000000 }
      responses:
        '202':
          description: 전송 작업의 현재 상태; 완료나 수신을 뜻하지 않음
          content:
            application/json:
              schema:
                type: object
                required: [operationId, status]
                properties:
                  operationId: { type: string }
                  status: { enum: [queued, smtp_accepted, sent_copy_pending, completed, rejected, delivery_unknown] }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: 같은 Idempotency-Key에 다른 요청 내용 }
        '413': { description: 첨부 또는 전체 메시지 용량 초과 }
        '429': { description: 요청 한도 초과, Retry-After 제공 }
  /api/settings/mail:
    get:
      summary: 실제 지원이 확인된 설정과 편집 가능 여부
      responses:
        '200':
          description: 지원 API를 확인하지 못한 기능의 editable은 false
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [key, editable, reason]
                  properties:
                    key: { enum: [signature, vacation, forwarding, filters, aliases] }
                    editable: { type: boolean }
                    reason: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/admin/mailboxes:
    get:
      summary: 서버에서 확인한 역할과 도메인 범위 내의 메일함 조회
      description: 메일 운영자는 허용 도메인만, 플랫폼 관리자는 명시된 범위만 조회. 생성·삭제 API는 별도 운영 설계 후 추가.
      parameters:
        - { name: domain, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: 원본 시스템·동기화 상태와 메일함 요약
          content:
            application/json:
              schema:
                type: object
                properties:
                  source: { const: mailcow }
                  syncedAt: { type: [string, 'null'], format: date-time }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        address: { type: string, format: email }
                        quotaBytes: { type: integer, minimum: 0 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-dyhs_session
      description: 불투명한 서버 세션. HttpOnly, Secure, SameSite=Lax, Path=/; Domain 속성 없음
  parameters:
    MessageId:
      name: id
      in: path
      required: true
      description: 서버가 메일함 소유권과 폴더+UIDVALIDITY+UID에 바인딩한 불투명 ID
      schema: { type: string, maxLength: 512 }
    Csrf:
      name: X-CSRF-Token
      in: header
      required: true
      description: 서버 세션에 바인딩한 토큰. Origin 검사도 필수
      schema: { type: string }
  schemas:
    MessageSummary:
      type: object
      required: [id, subject, from, receivedAt, unread]
      properties:
        id: { type: string }
        subject: { type: string }
        from: { type: string, format: email }
        receivedAt: { type: string, format: date-time }
        unread: { type: boolean }
        hasAttachments: { type: boolean }
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
  responses:
    Unauthorized:
      description: 로그인 필요 또는 세션 만료
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: SESSION_EXPIRED, message: 다시 로그인해 주세요., requestId: example-request }
    Forbidden:
      description: 역할·도메인·메일함 연결·CSRF 검증 실패. 대상 정보 비노출
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: ACCESS_DENIED, message: 이 작업을 수행할 권한이 없습니다., requestId: example-request }
    NotFound:
      description: 존재하지 않거나 세션 사용자 소유가 아닌 리소스
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: NOT_FOUND, message: 항목을 찾을 수 없습니다., requestId: example-request }
    Unavailable:
      description: 메일 서버 일시 연결 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: MAIL_UNAVAILABLE, message: 메일함에 연결하지 못했어요. 잠시 후 다시 시도해 주세요., requestId: example-request }